requirements.txt
torch>=2.0.0
open_clip_torch>=2.20.0
Pillow>=9.0.0
numpy>=1.21.0
streamlit>=1.28.0
streamlit-option-menu>=0.3.6
